Eth tests env setup extracted out in a helper function
